+++ src/rust/ide/src/ide.rs
@@ -1,40 +1,57 @@
 //! This module contains the IDE object implementation.
+pub mod initializer;
+pub mod integration;
 
 use crate::prelude::*;
 
-use crate::config;
-use crate::transport::web::WebSocket;
-use crate::view::View;
+use crate::controller::FilePath;
+use crate::model::module::Path as ModulePath;
+use crate::ide::integration::Integration;
 
-use enso_protocol::binary;
-use enso_protocol::traits::*;
-use enso_protocol::language_server;
-use enso_protocol::project_manager;
-use enso_protocol::project_manager::ProjectName;
-use uuid::Uuid;
-use crate::config::BackendService;
+use ensogl::application::Application;
+use ensogl::system::web;
+use enso_protocol::language_server::MethodPointer;
+use parser::Parser;
+
+pub use initializer::Initializer;
+
 
 
 // =================
 // === Constants ===
 // =================
 
-// TODO[ao] We need to set a big timeout on Project Manager to make sure it will have time to
-//          download required version of Engine. This should be handled properly when implementing
-//          https://github.com/enso-org/ide/issues/1034
-const PROJECT_MANAGER_TIMEOUT_SEC:u64 = 2 * 60 * 60;
+/// The name of the module initially opened in the project view.
+///
+/// Currently this name is hardcoded in the engine services and is populated for each project
+/// created using engine's Project Picker service.
+///
+/// TODO [mwu] Name of the module that will be initially opened in the text editor.
+///      Provisionally the Project View is hardcoded to open with a single text
+///      editor and it will be connected with a file with module of this name.
+///      To be replaced with better mechanism once we decide how to describe
+///      default initial layout for the project.
+pub const INITIAL_MODULE_NAME:&str = "Main";
 
+/// Name of the main definition.
+///
+/// This is the definition whose graph will be opened on IDE start.
+pub const MAIN_DEFINITION_NAME:&str = "main";
 
+/// The code with definition of the default `main` method.
+pub fn default_main_method_code() -> String {
+    format!(r#"{} = "Hello, World!""#, MAIN_DEFINITION_NAME)
+}
 
-// ==============
-// === Errors ===
-// ==============
+/// The default content of the newly created initial main module file.
+pub fn default_main_module_code() -> String {
+    default_main_method_code()
+}
 
-/// Error raised when project with given name was not found.
-#[derive(Clone,Debug,Fail)]
-#[fail(display="Project with nae {} was not found.", name)]
-pub struct ProjectNotFound {
-    name : ProjectName
+/// Method pointer that described the main method, i.e. the method that project view wants to open
+/// and which presence is currently required.
+pub fn main_method_ptr(project_name:impl Str, module_path:&model::module::Path) -> MethodPointer {
+    module_path.method_pointer(project_name,MAIN_DEFINITION_NAME)
 }
 
 
@@ -43,162 +60,114 @@ pub struct ProjectNotFound {
 // === Ide ===
 // ===========
 
-/// The IDE structure containing its configuration and its components instances.
+/// The main Ide structure.
+///
+/// This structure is a root of all objects in our application. It includes both layers:
+/// Controllers and Views, and an integration between them.
 #[derive(Debug)]
 pub struct Ide {
-    view : View
+    application : Application,
+    integration : Integration,
+}
+
+impl Ide {
+    /// Constructor.
+    pub async fn new(project:model::Project) -> FallibleResult<Self> {
+        let logger      = Logger::new("Ide");
+        let module_path = initial_module_path(&project)?;
+        let file_path   = module_path.file_path().clone();
+        // TODO [mwu] This solution to recreate missing main file should be considered provisional
+        //   until proper decision is made. See: https://github.com/enso-org/enso/issues/1050
+        recreate_if_missing(&project, &file_path, default_main_method_code()).await?;
+
+        let method = main_method_ptr(project.name(),&module_path);
+        let module = project.module(module_path).await?;
+        add_main_if_missing(&module,&method,&project.parser())?;
+
+        // Here, we should be relatively certain (except race conditions in case of multiple clients
+        // that we currently do not support) that main module exists and contains main method.
+        // Thus, we should be able to successfully create a graph controller for it.
+        let graph         = controller::ExecutedGraph::new(&logger,project.clone_ref(),method).await?;
+        let text          = controller::Text::new(&logger, &project, file_path).await?;
+        let visualization = project.visualization().clone();
+
+        let application   = Application::new(&web::get_html_element_by_id("root").unwrap());
+
+        let view = application.new_view::<ide_view::project::View>();
+        application.display.add_child(&view);
+
+        let integration = Integration::new(view,graph,text,visualization,project);
+        Ok(Ide {application,integration})
+    }
 }
 
 
-
-
-// ======================
-// === IdeInitializer ===
-// ======================
-
-/// The IDE initializer.
-#[derive(Debug)]
-pub struct IdeInitializer {
-    config : config::Startup,
-    logger : Logger
+/// Returns the path to the initially opened module in the given project.
+pub fn initial_module_path(project:&model::Project) -> FallibleResult<ModulePath> {
+    model::module::Path::from_name_segments(project.content_root_id(),&[INITIAL_MODULE_NAME])
 }
 
-
-impl IdeInitializer {
-    pub fn new_local() -> Self {
-        Self {
-            config : config::Startup::new_local().expect("Failed to load configuration."),
-            logger : Logger::new("IdeInitializer"),
-        }
+/// Create a file with default content if it does not already exist.
+pub async fn recreate_if_missing(project:&model::Project, path:&FilePath, default_content:String)
+-> FallibleResult {
+    let rpc = project.json_rpc();
+    if !rpc.file_exists(path).await?.exists {
+        rpc.write_file(path,&default_content).await?;
     }
-
-    pub fn start_and_forget(self) {
-        let executor = setup_global_executor();
-        executor::global::spawn(async move {
-            info!(self.logger, "Starting IDE with the following config: {self.config:?}");
-            // TODO [mwu] Once IDE gets some well-defined mechanism of reporting
-            //      issues to user, such information should be properly passed
-            //      in case of setup failure.
-            let project = self.initialize_project_model().await.expect("Failed to setup project model");
-            let view    = View::new(Logger::new("IDE"),project).await.expect("Failed to initialize project view");
-            info!(self.logger,"Setup done.");
-            let ide = Ide{view};
-            std::mem::forget(ide);
-        });
-        std::mem::forget(executor);
-    }
-
-    pub async fn initialize_project_model(&self) -> FallibleResult<model::Project>{
-        match &self.config.backend {
-            BackendService::ProjectManager { endpoint } => {
-                let project_manager = self.setup_project_manager(endpoint).await?;
-                let logger          = self.logger.clone_ref();
-                let project_name    = self.config.project_name.clone();
-                let initializer     = IdeWithProjectManagerInitializer{logger,project_manager,project_name};
-                initializer.open_project().await
-            }
-            BackendService::LanguageServer {json_endpoint,binary_endpoint} => {
-                create_project_model(&self.logger,None,json_endpoint.into(),binary_endpoint.into(),default(),self.config.project_name.clone()).await
-            }
-        }
-    }
-
-    /// Wraps the transport to the project manager server into the client type and registers it
-    /// within the global executor.
-    pub async fn setup_project_manager
-    (&self, endpoint:&str) -> FallibleResult<project_manager::Client> {
-        let transport           = WebSocket::new_opened(self.logger.clone_ref(),endpoint).await?;
-        let mut project_manager = project_manager::Client::new(transport);
-        project_manager.set_timeout(std::time::Duration::from_secs(PROJECT_MANAGER_TIMEOUT_SEC));
-        executor::global::spawn(project_manager.runner());
-        Ok(project_manager)
-    }
+    Ok(())
 }
 
-struct IdeWithProjectManagerInitializer {
-    logger          : Logger,
-    project_manager : project_manager::Client,
-    project_name    : ProjectName,
-}
-
-impl IdeWithProjectManagerInitializer {
-    /// Connect to language server.
-    pub async fn open_project(self) -> FallibleResult<model::Project> {
-        use project_manager::MissingComponentAction::*;
-
-        let project_id      = self.get_project_or_create_new().await?;
-        let endpoints       = self.project_manager.open_project(&project_id,&Install).await?;
-        let json_endpoint   = endpoints.language_server_json_address.to_string();
-        let binary_endpoint = endpoints.language_server_binary_address.to_string();
-        let project_manager:Rc<dyn project_manager::API> = Rc::new(self.project_manager);
-        let project_manager = Some(project_manager);
-        create_project_model(&self.logger,project_manager,json_endpoint,binary_endpoint,project_id,self.project_name).await
+/// Add main method definition to the given module, if the method is not already defined.
+///
+/// The lookup will be done using the given `main_ptr` value.
+pub fn add_main_if_missing
+(module:&model::Module, main_ptr:&MethodPointer, parser:&Parser) -> FallibleResult {
+    if module.lookup_method(main_ptr).is_err() {
+        let mut info  = module.info();
+        let main_code = default_main_method_code();
+        let main_ast  = parser.parse_line(main_code)?;
+        info.add_ast(main_ast,double_representation::module::Placement::End)?;
+        module.update_ast(info.ast)?;
     }
-
-    /// Creates a new project and returns its metadata, so the newly connected project can be
-    /// opened.
-    pub async fn create_project(&self) -> FallibleResult<Uuid> {
-        use project_manager::MissingComponentAction::Install;
-        info!(self.logger,"Creating a new project named '{self.project_name}'.");
-        let version           = None;
-        let ProjectName(name) = &self.project_name;
-        let response          = self.project_manager.create_project(name,&version,&Install);
-        Ok(response.await?.project_id)
-    }
-
-    async fn lookup_project(&self) -> FallibleResult<Uuid> {
-        let response     = self.project_manager.list_projects(&None).await?;
-        let mut projects = response.projects.iter();
-        projects.find(|project_metadata| {
-            project_metadata.name == self.project_name
-        }).map(|md| md.id).ok_or_else(|| ProjectNotFound{name:self.project_name.clone()}.into())
-    }
-
-    /// Returns project with `project_name` or returns a newly created one if it doesn't exist.
-    pub async fn get_project_or_create_new(&self) -> FallibleResult<Uuid> {
-        let project          = self.lookup_project().await;
-        if let Ok(project_id) = project {
-            Ok(project_id)
-        } else {
-            info!(self.logger, "Attempting to create {self.project_name}");
-            self.create_project().await
-        }
-    }
+    Ok(())
 }
 
 
 
 // =============
-// === Utils ===
+// === Tests ===
 // =============
 
-/// Creates a new running executor with its own event loop. Registers them as a global executor.
-pub fn setup_global_executor() -> executor::web::EventLoopExecutor {
-    let executor = executor::web::EventLoopExecutor::new_running();
-    executor::global::set_spawner(executor.spawner.clone());
-    executor
-}
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::executor::test_utils::TestWithLocalPoolExecutor;
 
-async fn create_project_model
-( logger : &Logger
-, project_manager : Option<Rc<dyn project_manager::API>>
-, json_endpoint   : String
-, binary_endpoint : String
-, project_id      : Uuid
-, project_name    : ProjectName
-) -> FallibleResult<model::Project> {
-    info!(logger, "Establishing Language Server connection.");
-    let client_id     = Uuid::new_v4();
-    let json_ws       = WebSocket::new_opened(logger,json_endpoint).await?;
-    let binary_ws     = WebSocket::new_opened(logger,binary_endpoint).await?;
-    let client_json   = language_server::Client::new(json_ws);
-    let client_binary = binary::Client::new(logger,binary_ws);
-    crate::executor::global::spawn(client_json.runner());
-    crate::executor::global::spawn(client_binary.runner());
-    let connection_json   = language_server::Connection::new(client_json,client_id).await?;
-    let connection_binary = binary::Connection::new(client_binary,client_id).await?;
-    let ProjectName(name) = project_name;
-    let project           = model::project::Synchronized::from_connections(logger,
-        project_manager,connection_json,connection_binary,project_id,name).await?;
-    Ok(Rc::new(project))
+    #[wasm_bindgen_test]
+    fn adding_missing_main() {
+        let _ctx        = TestWithLocalPoolExecutor::set_up();
+        let parser      = parser::Parser::new_or_panic();
+        let mut data    = crate::test::mock::Unified::new();
+        let module_name = data.module_path.module_name();
+        let main_ptr    = main_method_ptr(&data.project_name,&data.module_path);
+
+        // Check that module without main gets it after the call.
+        let empty_module_code = "";
+        data.set_code(empty_module_code);
+        let module = data.module();
+        assert!(module.lookup_method(&main_ptr).is_err());
+        add_main_if_missing(&module,&main_ptr,&parser).unwrap();
+        assert!(module.lookup_method(&main_ptr).is_ok());
+
+        // Now check that modules that have main already defined won't get modified.
+        let mut expect_intact = move |code:&str| {
+            data.set_code(code);
+            let module = data.module();
+            add_main_if_missing(&module,&main_ptr,&parser).unwrap();
+            assert_eq!(code,module.ast().repr());
+        };
+        expect_intact("main = 5");
+        expect_intact("here.main = 5");
+        expect_intact(&format!("{}.main = 5",module_name));
+    }
 }
